+++ crates/bevy_render/src/camera/projection.rs
@@ -1,9 +1,6 @@
-use super::{Camera, DepthCalculation};
-use bevy_ecs::Res;
-use bevy_math::{Mat4, Vec2, Vec3};
+use super::DepthCalculation;
+use bevy_math::Mat4;
 use bevy_reflect::{Reflect, ReflectComponent, ReflectDeserialize};
-use bevy_transform::components::GlobalTransform;
-use bevy_window::Windows;
 use serde::{Deserialize, Serialize};
 
 pub trait CameraProjection {
@@ -114,26 +111,4 @@ impl Default for OrthographicProjection {
             window_origin: WindowOrigin::Center,
         }
     }
-}
-
-/// Given coordinates in world space, use the camera and window information to compute the
-/// screen space coordinates.
-pub fn world_to_screen(
-    world_space_coords: &GlobalTransform,
-    camera: (&Camera, &GlobalTransform),
-    window_resource: &Res<Windows>,
-) -> Option<Vec2> {
-    let projection_matrix = camera.0.projection_matrix;
-    let window = window_resource.get(camera.0.window)?;
-    let window_size = Vec2::new(window.width(), window.height());
-    // Build a transform to convert from world to NDC using camera data
-    let world_to_ndc: Mat4 = projection_matrix * camera.1.compute_matrix().inverse();
-    let ndc_space_coords: Vec3 = world_to_ndc.transform_point3(world_space_coords.translation);
-    // NDC z-values outside of 0 < z < 1 are behind the camera and are thus not in screen space
-    if ndc_space_coords.z < 0.0 || ndc_space_coords.z > 1.0 {
-        return None;
-    }
-    // Once in NDC space, we can discard the z element and rescale x/y to fit the screen
-    let screen_space_coords = (ndc_space_coords.truncate() + Vec2::one()) / 2.0 * window_size;
-    Some(screen_space_coords)
 }
